delete redundant codes
