+++ src/pages/CalendarPage/CalendarPage.js
@@ -1,73 +1,28 @@
 import React, { useEffect, useState, useContext } from 'react';
 import styled from 'styled-components';
-import { useNavigate } from 'react-router-dom';
 
-//日曆
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 import moment from 'moment';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
 
-//firebase
-import { initializeApp } from 'firebase/app';
-import {
-  getFirestore,
-  doc,
-  setDoc,
-  collection,
-  onSnapshot,
-  query,
-  orderBy,
-  updateDoc,
-  deleteDoc,
-} from 'firebase/firestore';
+import { collection, onSnapshot, query, orderBy } from 'firebase/firestore';
+import { db } from '../../utils/firebase';
 
-//components
 import UserContext from '../../contexts/UserContext';
+import CalendarDetail from './CalendarDetail';
 
-//pic
-import trainingBanner from '../../images/Strong-man-doing-bench-press-in-gym-400451.jpg';
-import detailBanner from '../../images/Blue-gymnast-rings-1044340.jpg';
-import armMuscle from '../../images/armMuscle.png';
-
-//FontAwesomeIcon
-import { library } from '@fortawesome/fontawesome-svg-core';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCircleXmark, faClock, faFire, faDumbbell, faWeightHanging } from '@fortawesome/free-solid-svg-icons';
+import trainingBanner from '../../images/Strong-man-doing-bench-press-in-gym.jpg';
 
 function CalendarPage() {
-  //UserContext拿資料
-  const { isLoggedIn, setIsLoggedIn, userSignOut, signInWithGoogle, uid, displayName, email, signIn } =
-    useContext(UserContext);
+  const { isLoggedIn, uid } = useContext(UserContext);
 
-  //日曆
   const localizer = momentLocalizer(moment);
 
-  //抓菜單
   const [trainingData, setTrainingData] = useState([]);
   const [trainingItem, setTrainingItem] = useState([]);
 
-  //點行程抓資料
   const [detail, setDetail] = useState();
 
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  const firebaseConfig = {
-    apiKey: 'AIzaSyDtlWrSX2x1e0oTxI1_MN52sQsVyEwaOzA',
-    authDomain: 'fitness2-d4aaf.firebaseapp.com',
-    projectId: 'fitness2-d4aaf',
-    storageBucket: 'fitness2-d4aaf.appspot.com',
-    messagingSenderId: '440863323792',
-    appId: '1:440863323792:web:3f097801137f4002c7ca15',
-  };
-
-  const app = initializeApp(firebaseConfig);
-  const db = getFirestore(app);
-
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝即時抓出每筆菜單資料＝＝＝＝＝＝＝＝＝＝＝
-
-  //從firebase上即時抓資料
   useEffect(() => {
     async function getTrainingTables() {
       const docRef = query(collection(db, 'users', uid, 'trainingTables'), orderBy('trainingDate'));
@@ -82,7 +37,6 @@ function CalendarPage() {
     getTrainingTables();
   }, [isLoggedIn, uid]);
 
-  //map成日曆需要的格式
   useEffect(() => {
     if (trainingItem.length > 0) {
       const events = trainingItem.map((item, index) => {
@@ -103,8 +57,6 @@ function CalendarPage() {
 
   const events = trainingData;
 
-  // ＝＝＝＝＝＝＝＝＝＝即時抓出每筆菜單資料＝＝＝＝＝＝＝＝＝＝＝
-
   const NewHOC = (PassedComponent) => {
     return class extends React.Component {
       render() {
@@ -112,9 +64,6 @@ function CalendarPage() {
           <Calendar
             localizer={localizer}
             events={events}
-            // startAccessor="start"
-            // endAccessor="end"
-            // showMultiDayTimes
             style={{ minHeight: 900, background: '#F5F5F5', borderRadius: '12px', padding: '20px' }}
             components={{
               event: event,
@@ -127,258 +76,35 @@ function CalendarPage() {
 
   const event = ({ event }) => {
     return (
-      <Events
+      <div
         onClick={() => {
           setDetail(event);
         }}
       >
-        <EventName>{event.title} </EventName>
-        <EventDescription>{event.description}</EventDescription>
-      </Events>
+        <div>{event.title} </div>
+        <div>{event.description}</div>
+      </div>
     );
   };
 
   const NewComponent = NewHOC(event);
 
   return (
-    <Wrapper>
+    <>
       <BannerOutside>
         <Banner>
           <BannerText>讓健身成為生活的一部分！</BannerText>
         </Banner>
       </BannerOutside>
-      {detail ? (
-        <>
-          <DetailOutside>
-            <Close
-              onClick={() => {
-                setDetail(null);
-              }}
-            >
-              <FontAwesomeIcon icon={faCircleXmark} />
-            </Close>
-            <Top>
-              <DetailTitle>
-                <TitlePic>
-                  <FontAwesomeIcon icon={faFire} />
-                </TitlePic>
-                {detail.title}
-              </DetailTitle>
-              <Line />
-              <DetailActionsOutside>
-                {detail.actions.map((item, index) => (
-                  <ActionItem>
-                    <ActionBodyPart>
-                      <BodyPartPic src={armMuscle} />
-                      {item.bodyPart}
-                    </ActionBodyPart>
-                    <ActionName>
-                      <FaDumbbellName>
-                        <FontAwesomeIcon icon={faDumbbell} />
-                      </FaDumbbellName>
-                      {item.actionName}
-                    </ActionName>
-                    <Weight>
-                      <FaDumbbellWeight>
-                        <FontAwesomeIcon icon={faWeightHanging} />
-                      </FaDumbbellWeight>
-                      {item.weight}KG
-                    </Weight>
-                    <Times>
-                      <FaDumbbellTimes>
-                        <FontAwesomeIcon icon={faClock} />
-                      </FaDumbbellTimes>
-                      {item.times}次
-                    </Times>
-                  </ActionItem>
-                ))}
-              </DetailActionsOutside>
-            </Top>
-            <PicOutside>
-              <Pic />
-            </PicOutside>
-          </DetailOutside>{' '}
-          <Background $isActive={detail} />
-        </>
-      ) : null}
+      {detail && <CalendarDetail detail={detail} setDetail={setDetail} />}
       <CalendarOutside>
         <NewComponent />
       </CalendarOutside>
-    </Wrapper>
+    </>
   );
 }
 
 export default CalendarPage;
-
-const DetailOutside = styled.div`
-  margin: 0 auto;
-  position: absolute;
-  top: 13%;
-  left: calc(50% - 346px);
-  z-index: 15;
-  background: #475260;
-  max-width: 1000px;
-  color: white;
-  border-top: 0.5rem solid #74c6cc;
-  z-index: 20;
-  animation-name: detailfadein;
-  animation-duration: 0.5s;
-  @keyframes detailfadein {
-    0% {
-      transform: translateY(-2%);
-      opacity: 0%;
-    }
-    100% {
-      transform: translateY(0%);
-      opacity: 100%;
-    }
-  }
-  @media screen and (max-width: 767px) {
-    max-width: 320px;
-    left: calc(50% - 160px);
-  }
-`;
-
-const Close = styled.div`
-  cursor: pointer;
-  width: 30px;
-  position: absolute;
-  right: 20px;
-  top: 10px;
-  scale: 1;
-  transition: 0.3s;
-  font-size: 30px;
-  color: #c14e4f;
-  &:hover {
-    scale: 1.2;
-  }
-`;
-
-const Top = styled.div`
-  padding: 0px 20px;
-`;
-
-const TitlePic = styled.div`
-  margin-right: 12px;
-`;
-
-const DetailTitle = styled.div`
-  display: flex;
-  margin-top: 45px;
-  font-weight: 600;
-  letter-spacing: 3px;
-  color: #74c6cc;
-  font-size: 25px;
-`;
-
-const Line = styled.div`
-  border-bottom: 2px solid #74c6cc;
-  margin: 20px 0px;
-`;
-
-const DetailActionsOutside = styled.div`
-  font-size: 20px;
-`;
-
-const ActionItem = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  color: black;
-  border: 1px solid #818a8e;
-  margin: 10px 0px;
-  border: 1px solid #818a8e;
-  padding: 5px 10px 5px 10px;
-  background: rgba(255, 255, 255, 0.5);
-  @media screen and (max-width: 1279px) {
-    flex-wrap: wrap;
-  }
-`;
-
-const BodyPartPic = styled.img`
-  object: fit;
-  width: 25px;
-  margin-right: 10px;
-  @media screen and (max-width: 767px) {
-    width: 25px;
-    margin-right: 10px;
-  }
-`;
-
-const ActionBodyPart = styled.div`
-  width: 150px;
-  display: flex;
-  @media screen and (max-width: 767px) {
-    width: 200px;
-    margin: 5px 0px;
-  }
-`;
-
-const FaDumbbellName = styled.div`
-  margin-right: 10px;
-  color: white;
-`;
-
-const ActionName = styled.div`
-  width: 250px;
-  display: flex;
-  @media screen and (max-width: 767px) {
-    margin: 5px 0px;
-  }
-`;
-
-const FaDumbbellWeight = styled.div`
-  margin-right: 14px;
-  color: white;
-`;
-
-const Weight = styled.div`
-  width: 150px;
-  display: flex;
-  @media screen and (max-width: 767px) {
-    width: 200px;
-    margin: 5px 0px;
-  }
-`;
-
-const FaDumbbellTimes = styled.div`
-  margin-right: 13px;
-  color: white;
-`;
-
-const Times = styled.div`
-  width: 80px;
-  display: flex;
-  margin: 5px 0px;
-`;
-
-const PicOutside = styled.div`
-  width: 100%;
-  height: 200px;
-  margin: 30px 0px 40px 0px;
-`;
-
-const Pic = styled.div`
-  width: 100%;
-  height: 100%;
-  background-image: url(${detailBanner});
-  background-size: cover;
-  background-position: 30% 50%;
-  background-repeat: no-repeat;
-`;
-
-const Background = styled.div`
-  background: black;
-  top: 0;
-  opacity: 50%;
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  z-index: 10;
-  display: ${(props) => (props.$isActive ? 'block;' : 'none;')};
-`;
-
-const Wrapper = styled.div``;
 
 const BannerOutside = styled.div`
   height: 320px;
@@ -432,9 +158,3 @@ const CalendarOutside = styled.div`
     padding: 80px 80px 80px 80px;
   }
 `;
-
-const Events = styled.div``;
-
-const EventName = styled.div``;
-
-const EventDescription = styled.div``;
