+++ src/pages/Training/FavoritePage.js
@@ -1,100 +1,27 @@
-import React, { useContext, useEffect, useState, useRef } from 'react';
+import React, { useContext, useState } from 'react';
 import styled from 'styled-components';
 
-//components
 import UserContext from '../../contexts/UserContext';
+import FavoriteNameZone from './FavoriteNameZone';
+import FavoriteActionZone from './FavoriteActionZone';
 
-//beautiful-dnd
-import { Draggable, DragDropContext, Droppable } from 'react-beautiful-dnd';
+import { doc, updateDoc, deleteDoc } from 'firebase/firestore';
+import { db } from '../../utils/firebase';
 
-//firebase
-import { initializeApp } from 'firebase/app';
-import {
-  getFirestore,
-  doc,
-  setDoc,
-  collection,
-  onSnapshot,
-  query,
-  orderBy,
-  updateDoc,
-  deleteDoc,
-  getDoc,
-} from 'firebase/firestore';
-import { getStorage, ref, uploadBytes, getDownloadURL } from 'firebase/storage';
-
-//FontAwesomeIcon
-import { library } from '@fortawesome/fontawesome-svg-core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faCircleXmark,
-  faTrash,
-  faDumbbell,
-  faPenToSquare,
-  faXmark,
-  faHeartCirclePlus,
-  faTriangleExclamation,
-} from '@fortawesome/free-solid-svg-icons';
-import {} from '@fortawesome/free-brands-svg-icons';
+import { faCircleXmark, faHeartCirclePlus, faTriangleExclamation } from '@fortawesome/free-solid-svg-icons';
 
-//pic
-import favoriteBanner from '../../images/Kettlebell-weights-724190.jpg';
-import armMuscle from '../../images/armMuscle.png';
+import favoriteBanner from '../../images/Kettlebell-weights.jpg';
 
 const FavoritePage = (props) => {
-  //UserContext拿資料
-  const {
-    isLoggedIn,
-    setIsLoggedIn,
-    userSignOut,
-    signInWithGoogle,
-    uid,
-    displayName,
-    email,
-    signIn,
-    alertPop,
-    setContent,
-  } = useContext(UserContext);
+  const { uid, alertPop, setContent } = useContext(UserContext);
 
-  //編輯名稱
   const [pickName, setPickName] = useState(null);
   const [newName, setNewName] = useState('');
 
-  //刪除功能
   const [deleteAlert, setDeleteAlert] = useState(false);
   const [deleteItem, setDeleteItem] = useState();
 
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  const firebaseConfig = {
-    apiKey: 'AIzaSyDtlWrSX2x1e0oTxI1_MN52sQsVyEwaOzA',
-    authDomain: 'fitness2-d4aaf.firebaseapp.com',
-    projectId: 'fitness2-d4aaf',
-    storageBucket: 'fitness2-d4aaf.appspot.com',
-    messagingSenderId: '440863323792',
-    appId: '1:440863323792:web:3f097801137f4002c7ca15',
-  };
-
-  const app = initializeApp(firebaseConfig);
-  const db = getFirestore(app);
-  const storage = getStorage(app);
-
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  //點選主題
-  function changePick(index) {
-    props.setPickFavorite(index);
-    props.setPickActions(props.favoriteTrainings[index].actions);
-    props.setPickID(props.favoriteTrainings[index].docID);
-  }
-
-  //刪除該主題通知跳出
-  async function deletePick(index) {
-    setDeleteAlert(true);
-    setDeleteItem(index);
-  }
-
-  //刪除該主題
   async function confirmDeletePick(index) {
     try {
       const docRef = await doc(db, 'users', uid, 'favoriteTrainings', props.favoriteTrainings[index].docID);
@@ -108,7 +35,6 @@ const FavoritePage = (props) => {
     }
   }
 
-  //Dnd改變順序
   const onDragEnd = (result) => {
     if (!result.destination) return;
     const items = Array.from(props.pickActions);
@@ -127,13 +53,11 @@ const FavoritePage = (props) => {
     updateActions();
   };
 
-  //改變名字
   function changeName(index) {
     setPickName(index);
     setNewName(props.favoriteTrainings[index].title);
   }
 
-  //送出新名字
   async function updateNewName(index) {
     const docRef = await doc(db, 'users', uid, 'favoriteTrainings', props.pickID);
     const data = {
@@ -177,156 +101,63 @@ const FavoritePage = (props) => {
           <DeleteBackground />
         </>
       )}
-      <Wrapper $isActive={props.openFavorite}>
-        <Close
-          onClick={() => {
-            props.setOpenFavorite(false);
-            props.setPickFavorite(null);
-            props.setPickActions([]);
-            setPickName(null);
-          }}
-        >
-          <FontAwesomeIcon icon={faCircleXmark} />
-        </Close>
-        <Top>
-          <Title>
-            <FaHeartCirclePlus>
-              <FontAwesomeIcon icon={faHeartCirclePlus} />
-            </FaHeartCirclePlus>
-            喜愛菜單
-          </Title>
-          <Line />
-        </Top>
-        <Bottom>
-          {props.favoriteTrainings.length > 0 ? (
-            <>
-              <NameZone>
-                <Name> 主題</Name>
-                <NameContent>
-                  {props.favoriteTrainings.map((item, index) => (
-                    <NameContentListOustide
-                      index={index}
-                      onClick={() => {
-                        changePick(index);
-                      }}
-                      $isActive={index == props.pickFavorite}
-                    >
-                      <NameTitle>
-                        <FaXmark
-                          index={index}
-                          $isClick={index == pickName}
-                          onClick={() => {
-                            setPickName(null);
-                            setNewName('');
-                          }}
-                        >
-                          <FontAwesomeIcon icon={faXmark} style={{ pointerEvents: 'none' }} />
-                        </FaXmark>
-                        <FaPenToSquare
-                          index={index}
-                          $isClick={index !== pickName}
-                          onClick={() => {
-                            changeName(index);
-                          }}
-                        >
-                          <FontAwesomeIcon icon={faPenToSquare} style={{ pointerEvents: 'none' }} />
-                        </FaPenToSquare>
-                        <NameOld $isClick={index !== pickName}> {item.title}</NameOld>
-                        <NameNew
-                          index={index}
-                          $isClick={index == pickName}
-                          onChange={(e) => setNewName(e.target.value)}
-                          defaultValue={item.title}
-                          maxLength={10}
-                        />
-                        <UpdateName
-                          index={index}
-                          $isClick={index == pickName}
-                          onClick={() => {
-                            updateNewName(index);
-                          }}
-                        >
-                          送出
-                        </UpdateName>
-                      </NameTitle>
-                      <NameDelete
-                        onClick={() => {
-                          deletePick(index);
-                        }}
-                      >
-                        <FontAwesomeIcon icon={faTrash} style={{ pointerEvents: 'none' }} />
-                      </NameDelete>
-                    </NameContentListOustide>
-                  ))}
-                  <NameRemind>＊修改最多輸入10字</NameRemind>
-                </NameContent>
-              </NameZone>
-              <ActionZone>
-                <Action>動作</Action>
-                <ActionContent>
-                  {props.pickActions.length > 0 ? (
-                    <DragDropContext onDragEnd={onDragEnd}>
-                      <Droppable droppableId="list">
-                        {(provided) => (
-                          <div ref={provided.innerRef} {...provided.droppableProps}>
-                            {props.pickActions.map((item, index) => (
-                              <Draggable key={item.id} draggableId={item.id.toString()} index={index}>
-                                {(provided, snapshot) => (
-                                  <ActionContentListOustide
-                                    index={index}
-                                    {...provided.draggableProps}
-                                    {...provided.dragHandleProps}
-                                    ref={provided.innerRef}
-                                    style={{
-                                      ...provided.draggableProps.style,
-                                      border: snapshot.isDragging ? '2px solid #74c6cc' : 'none',
-                                      borderStyle: snapshot.isDragging ? 'outset' : 'none',
-                                      background: snapshot.isDragging ? '#74c6cc' : 'rgba(255, 255, 255, 0.5)',
-                                    }}
-                                  >
-                                    <ActionPart>
-                                      <BodyPartPic src={armMuscle} />
-                                      {item.bodyPart}
-                                    </ActionPart>
-                                    <ActionName>
-                                      <FaDumbbellName>
-                                        <FontAwesomeIcon icon={faDumbbell} />
-                                      </FaDumbbellName>
-                                      {item.actionName}
-                                    </ActionName>
-                                  </ActionContentListOustide>
-                                )}
-                              </Draggable>
-                            ))}
-                            {provided.placeholder}
-                          </div>
-                        )}
-                      </Droppable>
-                    </DragDropContext>
-                  ) : (
-                    <>
-                      <NoAction>
-                        請點選查看主題
-                        <br />
-                        並上下拖曳改變順序
-                      </NoAction>
-                    </>
-                  )}
-                </ActionContent>
-              </ActionZone>
-            </>
-          ) : (
-            <NoFavorite>尚未建立喜愛菜單</NoFavorite>
-          )}
-        </Bottom>
-        <PicOutside>
-          <Pic />
-        </PicOutside>
-      </Wrapper>
-      <Background $isActive={props.openFavorite} />
+      {props.openFavorite && (
+        <>
+          <Wrapper>
+            <Close
+              onClick={() => {
+                props.setOpenFavorite(false);
+                props.setPickFavorite(null);
+                props.setPickActions([]);
+                setPickName(null);
+              }}
+            >
+              <FontAwesomeIcon icon={faCircleXmark} />
+            </Close>
+            <Top>
+              <Title>
+                <FaHeartCirclePlus>
+                  <FontAwesomeIcon icon={faHeartCirclePlus} />
+                </FaHeartCirclePlus>
+                喜愛菜單
+              </Title>
+              <Line />
+            </Top>
+            <Bottom>
+              {props.favoriteTrainings.length > 0 ? (
+                <>
+                  <FavoriteNameZone
+                    pickName={pickName}
+                    setPickName={setPickName}
+                    changeName={changeName}
+                    setNewName={setNewName}
+                    updateNewName={updateNewName}
+                    setDeleteAlert={setDeleteAlert}
+                    setDeleteItem={setDeleteItem}
+                    favoriteTrainings={props.favoriteTrainings}
+                    pickFavorite={props.pickFavorite}
+                    setPickFavorite={props.setPickFavorite}
+                    setPickActions={props.setPickActions}
+                    setPickID={props.setPickID}
+                  />
+                  <FavoriteActionZone onDragEnd={onDragEnd} pickActions={props.pickActions} />
+                </>
+              ) : (
+                <NoFavorite>尚未建立喜愛菜單</NoFavorite>
+              )}
+            </Bottom>
+            <PicOutside>
+              <Pic />
+            </PicOutside>
+          </Wrapper>
+          <Background />
+        </>
+      )}
     </>
   );
 };
+
+export default FavoritePage;
 
 const DeleteAlertOutside = styled.div`
   display: flex;
@@ -415,31 +246,14 @@ const Yes = styled.div`
   font-weight: 600;
 `;
 
-const NoOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 50px;
-  color: black;
-  cursor: pointer;
-  border-radius: 10px;
+const NoOutside = styled(YesOutside)`
+  margin-right: 0px;
   &:hover {
     background: white;
-    color: black;
-  }
-  @media screen and (max-width: 767px) {
-    width: 120px;
-    margin: 40px 20px 40px 0px;
   }
 `;
 
-const No = styled.div`
-  padding: 3px;
-  font-size: 20px;
-  letter-spacing: 2px;
-  font-weight: 600;
-`;
+const No = styled(Yes)``;
 
 const DeleteBackground = styled.div`
   background: black;
@@ -452,7 +266,7 @@ const DeleteBackground = styled.div`
 `;
 
 const Wrapper = styled.div`
-  display: ${(props) => (props.$isActive ? 'block;' : 'none;')};
+  display: block;
   margin: 0 auto;
   position: absolute;
   top: 15%;
@@ -551,205 +365,6 @@ const Bottom = styled.div`
   }
 `;
 
-const NameZone = styled.div`
-  margin-right: 40px;
-  width: 100%;
-`;
-
-const Name = styled.div`
-  font-size: 25px;
-  letter-spacing: 12px;
-`;
-
-const NameContent = styled.div`
-  height: 327px;
-  overflow-y: scroll;
-  padding-right: 15px;
-  &::-webkit-scrollbar {
-    width: 6px;
-  }
-  &::-webkit-scrollbar-button {
-    display: none;
-  }
-  &::-webkit-scrollbar-track-piece {
-    background: transparent;
-  }
-  &::-webkit-scrollbar-thumb {
-    border-radius: 4px;
-    background-color: rgba(0, 0, 0, 0.4);
-    border: 1px solid slategrey;
-  }
-  &::-webkit-scrollbar-track {
-    box-shadow: transparent;
-  }
-  @media screen and (max-width: 767px) {
-    margin-bottom: 50px;
-  }
-`;
-
-const NameRemind = styled.div`
-  color: #cd5c5c;
-  font-size: 16px;
-  letter-spacing: 2px;
-  margin-top: -6px;
-`;
-
-const NameContentListOustide = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 300px;
-  margin: 15px 0px;
-  cursor: pointer;
-  color: black;
-  background: ${(props) => (props.$isActive ? 'rgb(116, 198, 204)' : 'rgba(255, 255, 255, 0.5)')};
-  padding: 5px 15px 5px 15px;
-  &:hover {
-    background: #74c6cc;
-  }
-  @media screen and (max-width: 767px) {
-    width: 100%;
-  }
-`;
-
-const NameTitle = styled.div`
-  display: flex;
-`;
-
-const FaPenToSquare = styled.div`
-  display: ${(props) => (props.$isClick ? 'block' : 'none')};
-  margin-right: 15px;
-  &:hover {
-    color: white;
-  }
-`;
-
-const FaXmark = styled.div`
-  display: ${(props) => (props.$isClick ? 'block' : 'none')};
-  margin-right: 15px;
-  &:hover {
-    color: white;
-  }
-`;
-
-const NameOld = styled.div`
-  display: ${(props) => (props.$isClick ? 'block' : 'none')};
-`;
-
-const NameNew = styled.input`
-  display: ${(props) => (props.$isClick ? 'block' : 'none')};
-  width: 80px;
-  font-size: 20px;
-`;
-
-const UpdateName = styled.div`
-  display: ${(props) => (props.$isClick ? 'block' : 'none')};
-  margin-left: 15px;
-  border-radius: 12px;
-  background: white;
-  padding: 0px 5px;
-  &:hover {
-    color: white;
-    background: black;
-  }
-`;
-
-const NameDelete = styled.div`
-  cursor: pointer;
-  &:hover {
-    color: red;
-  }
-`;
-
-const ActionZone = styled.div``;
-
-const BodyPartPic = styled.img`
-  object: fit;
-  width: 25px;
-  margin-right: 10px;
-  @media screen and (max-width: 767px) {
-    width: 25px;
-    margin-right: 10px;
-  }
-`;
-
-const ActionContent = styled.div`
-  height: 327px;
-  width: 365px;
-  overflow-y: scroll;
-  padding-right: 15px;
-  &::-webkit-scrollbar {
-    width: 6px;
-  }
-  &::-webkit-scrollbar-button {
-    display: none;
-  }
-  &::-webkit-scrollbar-track-piece {
-    background: transparent;
-  }
-  &::-webkit-scrollbar-thumb {
-    border-radius: 4px;
-    background-color: rgba(0, 0, 0, 0.4);
-    border: 1px solid slategrey;
-  }
-  &::-webkit-scrollbar-track {
-    box-shadow: transparent;
-  }
-  @media screen and (max-width: 500px) {
-    width: 280px;
-    padding-right: 0px;
-  }
-`;
-
-const ActionContentListOustide = styled.div`
-  display: flex;
-  justify-content: start;
-  align-items: center;
-  margin: 15px 0px;
-  cursor: pointer;
-  color: black;
-  background: rgba(255, 255, 255, 0.5);
-  padding: 5px 15px 5px 15px;
-  @media screen and (max-width: 500px) {
-    flex-direction: column;
-    align-items: start;
-    width: 95%;
-  }
-`;
-
-const Action = styled.div`
-  font-size: 25px;
-  letter-spacing: 12px;
-`;
-
-const ActionPart = styled.div`
-  width: 100px;
-`;
-
-const FaDumbbellName = styled.div`
-  color: white;
-  margin-right: 10px;
-`;
-
-const ActionName = styled.div`
-  display: flex;
-`;
-
-const NoAction = styled.div`
-  text-align: center;
-  line-height: 35px;
-  width: 320px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 280px;
-  border: 1px solid #818a8e;
-  margin-top: 12px;
-  @media screen and (max-width: 500px) {
-    width: 240px;
-  }
-`;
-
 const NoFavorite = styled.div`
   width: 720px;
   display: flex;
@@ -774,7 +389,7 @@ const Background = styled.div`
   position: fixed;
   width: 100vw;
   height: 100vh;
-  display: ${(props) => (props.$isActive ? 'block;' : 'none;')};
+  display: block;
 `;
 
 const PicOutside = styled.div`
@@ -791,5 +406,3 @@ const Pic = styled.div`
   background-position: 30% 75%;
   background-repeat: no-repeat;
 `;
-
-export default FavoritePage;
