+++ src/pages/Training/Training.js
@@ -1,10 +1,7 @@
 import React, { useContext, useEffect, useState, useRef } from 'react';
 import styled from 'styled-components';
 
-//firebase
-import { initializeApp } from 'firebase/app';
 import {
-  getFirestore,
   doc,
   setDoc,
   collection,
@@ -18,80 +15,50 @@ import {
   startAfter,
   limit,
 } from 'firebase/firestore';
-import { getStorage, ref, uploadBytes, getDownloadURL } from 'firebase/storage';
+import { ref, uploadBytes, getDownloadURL } from 'firebase/storage';
+import { db, storage } from '../../utils/firebase';
 
-//components
 import UserContext from '../../contexts/UserContext';
 import HistoryZone from './HistoryZone';
 import OpenHistoryZone from './OpenHistoryZone';
-import ChoiceActionOutsideZone from './ChoiceActionOutsideZone';
-import PromoteActionOutsideZone from './PromoteActionOutsideZone';
-import CalculationShowZone from './CalculationShowZone';
 import FavoritePage from './FavoritePage';
 import SkeletonPage from './SkeletonPage';
+import TrainingTwo from './TrainingTwo';
+import DeleteZone from './DeleteZone';
+import TrainingButtonsZone from './TrainingButtonsZone';
 
-//引入動作菜單
 import ACTIONS from './allActionsLists';
 
-//寄信
 import emailjs from 'emailjs-com';
 
-//chart.js
 import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
 
-//pic
 import trainingBanner from '../../images/Beautiful-woman-holding-heavy-604970.jpg';
-import pageOnePic from '../../images/Empty-gym-in-sunlight-397510.jpg';
-import logo from '../../images/高畫質logo_藍色2.png';
+import pageOnePic from '../../images/Empty-gym-in-sunlight.jpg';
 
-//FontAwesomeIcon
-import { library } from '@fortawesome/fontawesome-svg-core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faCircleArrowRight,
-  faCircleArrowLeft,
   faCircleXmark,
   faCalendarDays,
   faHandPointUp,
   faDumbbell,
-  faTriangleExclamation,
 } from '@fortawesome/free-solid-svg-icons';
-import {} from '@fortawesome/free-brands-svg-icons';
 
-//loading animation
 import { Blocks } from 'react-loader-spinner';
 
-//uuid
 import { v4 as uuid } from 'uuid';
 
 const Training = () => {
-  //UserContext拿資料
-  const {
-    isLoggedIn,
-    setIsLoggedIn,
-    userSignOut,
-    signInWithGoogle,
-    uid,
-    displayName,
-    email,
-    signIn,
-    alertPop,
-    setContent,
-  } = useContext(UserContext);
+  const { isLoggedIn, uid, displayName, email, signIn, alertPop, setContent } = useContext(UserContext);
 
-  //抓到每筆菜單
   const [trainingData, setTrainingData] = useState([]);
   const [pagination, setPagination] = useState([]);
   const [currentPage, setCurrentPgae] = useState(1);
 
-  //建立菜單上下頁開關
   const [openTrainingInput, setOpenTrainingInput] = useState(false);
-  const [openTrainingOne, setOpenTrainingOne] = useState(false);
-  const [openTrainingTwo, setOpenTrainingTwo] = useState(false);
-  const [openCompleteSetting, setOpenCompleteSetting] = useState(false);
-  const [pageTwo, setPageTwo] = useState(false);
+  const [openTrainingPage, setOpenTrainingPage] = useState(0);
 
-  //抓到菜單input
   const [title, setTitle] = useState('');
   const [date, setDate] = useState('');
   const [description, setDescription] = useState('');
@@ -99,75 +66,47 @@ const Training = () => {
   const [promoteActions, setPromoteActions] = useState([]);
   const [choiceAction, setChoiceAction] = useState([]);
 
-  //計算總重量
-  const [totalWeight, setTotalWeight] = useState(0);
-  const [choiceWeight, setChoiceWeight] = useState(0);
-  const [choiceTimes, setChoiceTimes] = useState(0);
-
-  //身體部位佔比
-  const [shoulderPercent, setShoulderPercent] = useState(0);
-  const [armPercent, setArmPercent] = useState(0);
-  const [chestPercent, setChestPercent] = useState(0);
-  const [backPercent, setBackPercent] = useState(0);
-  const [buttLegPercent, setButtLegPercent] = useState(0);
-  const [corePercent, setCorePercent] = useState(0);
-
-  //點擊哪個菜單就顯示哪個菜單
-  const [showHistory, setShowHistory] = useState([]);
+  const [showHistory, setShowHistory] = useState();
   const [showHistoryToggle, setShowHistoryToggle] = useState(false);
   const [showHistoryActions, setShowHistoryActions] = useState([]);
-  const [showHistoryBackground, setShowHistoryBackground] = useState(false);
   const [historyIndex, setHistoryIndex] = useState();
 
-  //上傳照片
   const [imageUpload, setImageUpload] = useState();
   const [imageList, setImageList] = useState('');
   const [pickHistory, setPickHistory] = useState();
-  const [showPicture, setShowPicture] = useState(true);
 
-  //點擊顯示影片
   const [videoUrl, setVideoUrl] = useState('');
   const [videoShow, setVideoShow] = useState(false);
-  const [playing, setPlaying] = useState();
+  const [playing, setPlaying] = useState(null);
 
-  //loading動畫
   const [loading, setLoading] = useState(false);
   const [skeleton, setSkeleton] = useState(false);
   const [uploadSkeleton, setUploadSkeleton] = useState(false);
 
-  //喜愛動作
   const [favoriteTrainings, setFavoriteTrainings] = useState([]);
   const [favoriteChoice, setFavoriteChoice] = useState(null);
 
-  //打開喜愛菜單
   const [openFavorite, setOpenFavorite] = useState(false);
   const [pickActions, setPickActions] = useState([]);
   const [pickFavorite, setPickFavorite] = useState(null);
   const [pickID, setPickID] = useState();
 
-  //確認刪除
   const [deleteAlert, setDeleteAlert] = useState(false);
 
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  const firebaseConfig = {
-    apiKey: 'AIzaSyDtlWrSX2x1e0oTxI1_MN52sQsVyEwaOzA',
-    authDomain: 'fitness2-d4aaf.firebaseapp.com',
-    projectId: 'fitness2-d4aaf',
-    storageBucket: 'fitness2-d4aaf.appspot.com',
-    messagingSenderId: '440863323792',
-    appId: '1:440863323792:web:3f097801137f4002c7ca15',
-  };
-
-  const app = initializeApp(firebaseConfig);
-  const db = getFirestore(app);
-  const storage = getStorage(app);
-
-  // ＝＝＝＝＝＝＝＝＝＝＝啟動firebase＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝＝chart.js＝＝＝＝＝＝＝＝＝＝＝
-
   ChartJS.register(ArcElement, Tooltip, Legend);
+
+  const shoulderNumber = choiceAction.filter((item) => item.bodyPart === '肩').length;
+  const armNumber = choiceAction.filter((item) => item.bodyPart === '手臂').length;
+  const chestNumber = choiceAction.filter((item) => item.bodyPart === '胸').length;
+  const backNumber = choiceAction.filter((item) => item.bodyPart === '背').length;
+  const buttLegNumber = choiceAction.filter((item) => item.bodyPart === '臀腿').length;
+  const coreNumber = choiceAction.filter((item) => item.bodyPart === '核心').length;
+  const shoulderPercent = shoulderNumber / choiceAction.length;
+  const armPercent = armNumber / choiceAction.length;
+  const chestPercent = chestNumber / choiceAction.length;
+  const backPercent = backNumber / choiceAction.length;
+  const buttLegPercent = buttLegNumber / choiceAction.length;
+  const corePercent = coreNumber / choiceAction.length;
 
   const data = {
     datasets: [
@@ -200,51 +139,29 @@ const Training = () => {
     labels: ['無資料'],
   };
 
-  useEffect(() => {
-    const shoulderNumber = choiceAction.filter((item) => item.bodyPart == '肩').length;
-    const armNumber = choiceAction.filter((item) => item.bodyPart == '手臂').length;
-    const chestNumber = choiceAction.filter((item) => item.bodyPart == '胸').length;
-    const backNumber = choiceAction.filter((item) => item.bodyPart == '背').length;
-    const buttLegNumber = choiceAction.filter((item) => item.bodyPart == '臀腿').length;
-    const coreNumber = choiceAction.filter((item) => item.bodyPart == '核心').length;
-    setShoulderPercent(shoulderNumber / choiceAction.length);
-    setArmPercent(armNumber / choiceAction.length);
-    setChestPercent(chestNumber / choiceAction.length);
-    setBackPercent(backNumber / choiceAction.length);
-    setButtLegPercent(buttLegNumber / choiceAction.length);
-    setCorePercent(coreNumber / choiceAction.length);
-  });
+  let re = /^[0-9]+.?[0-9]*$/;
+  let totalWeight;
+  const total = choiceAction.reduce((prev, item) => prev + item.weight * item.times, 0);
+  if (!re.test(total)) {
+    alertPop();
+    setContent('請輸入數字');
+  } else {
+    totalWeight = Number(total.toFixed(1));
+  }
 
-  // ＝＝＝＝＝＝＝＝＝＝＝chart.js＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝點擊建立菜單+上下頁切換＝＝＝＝＝＝＝＝＝＝＝
-
-  //點擊建立菜單
   function addTraining() {
     if (isLoggedIn) {
-      setShowHistoryBackground(true);
       setOpenTrainingInput(true);
-      setPageTwo(false);
-      setShowHistory([]);
-      if (openTrainingOne !== true && openTrainingTwo !== true && openCompleteSetting !== true)
-        setOpenTrainingOne(true);
+      setOpenTrainingPage(1);
       setImageList('');
     } else {
       signIn();
     }
   }
 
-  function getPageOne() {
-    setOpenTrainingOne(true);
-    setOpenTrainingTwo(false);
-    setPageTwo(false);
-  }
-
   function getPageTwo() {
     if (title !== '' && date !== '' && description !== '') {
-      setOpenTrainingOne(false);
-      setOpenTrainingTwo(true);
-      setPageTwo(true);
+      setOpenTrainingPage(2);
       setPart('肩');
     } else {
       alertPop();
@@ -252,28 +169,23 @@ const Training = () => {
     }
   }
 
-  //點擊右上角XX
   function closeAddTraining() {
     setOpenTrainingInput(false);
-    setOpenTrainingOne(false);
-    setOpenTrainingTwo(false);
-    setOpenCompleteSetting(false);
-    setShowHistoryBackground(false);
+    setOpenTrainingPage(0);
     setTitle('');
     setDate('');
     setDescription('');
     setChoiceAction([]);
-    setTotalWeight(0);
+    setVideoShow(false);
+    setPlaying(null);
   }
 
-  //點擊「完成本次鍛鍊」
   function completeTraining() {
     alertPop();
     setContent('恭喜完成本次鍛鍊');
     changeCompleteCondition();
   }
 
-  //改變狀態
   async function changeCompleteCondition() {
     try {
       const docRef = await doc(db, 'users', uid, 'trainingTables', pickHistory);
@@ -290,19 +202,12 @@ const Training = () => {
     }
   }
 
-  //可以刪除菜單
-  async function deleteTrainingItem() {
-    setDeleteAlert(true);
-  }
-
-  //真的刪除
   async function confirmDeleteTrainingItem() {
     setLoading(true);
     try {
       const docRef = await doc(db, 'users', uid, 'trainingTables', pickHistory);
       await deleteDoc(docRef);
       setShowHistoryToggle(false);
-      setShowHistoryBackground(false);
       setLoading(false);
       alertPop();
       setContent('成功刪除菜單');
@@ -313,29 +218,26 @@ const Training = () => {
     }
   }
 
-  // ＝＝＝＝＝＝＝＝＝＝點擊建立菜單+上下頁切換＝＝＝＝＝＝＝＝＝＝＝
-
   useEffect(() => {
-    if (part == '背') {
+    if (part === '背') {
       setPromoteActions(ACTIONS.BACK);
-    } else if (part == '手臂') {
+    } else if (part === '手臂') {
       setPromoteActions(ACTIONS.ARM);
-    } else if (part == '胸') {
+    } else if (part === '胸') {
       setPromoteActions(ACTIONS.CHEST);
-    } else if (part == '臀腿') {
+    } else if (part === '臀腿') {
       setPromoteActions(ACTIONS.BUTTLEG);
-    } else if (part == '核心') {
+    } else if (part === '核心') {
       setPromoteActions(ACTIONS.CORE);
-    } else if (part == '肩') {
+    } else if (part === '肩') {
       setPromoteActions(ACTIONS.SHOULDER);
-    } else if (part == '上半身') {
+    } else if (part === '上半身') {
       setPromoteActions(ACTIONS.UPPERBODY);
-    } else if (part == '全身') {
+    } else if (part === '全身') {
       setPromoteActions(ACTIONS.ALL);
     }
   }, [part]);
 
-  //從右邊加入左邊
   function addActionItem(e) {
     setChoiceAction((pre) => {
       const addAction = { ...promoteActions[e.target.id] };
@@ -349,7 +251,6 @@ const Training = () => {
     });
   }
 
-  //左邊的可以刪除
   function deleteItem(id) {
     const newNextChoiceAction = choiceAction.filter((item, index) => {
       return index !== id;
@@ -357,27 +258,10 @@ const Training = () => {
     setChoiceAction(newNextChoiceAction);
   }
 
-  //加總每個動作的重量
-  useEffect(() => {
-    let re = /^[0-9]+.?[0-9]*$/;
-    const total = choiceAction.reduce((prev, item) => prev + item.weight * item.times, 0);
-    if (!re.test(total)) {
-      alertPop();
-      setContent('請輸入數字');
-    } else {
-      setTotalWeight(Number(total.toFixed(1)));
-    }
-  }, [choiceAction, choiceWeight, choiceTimes]);
-
-  // ＝＝＝＝＝＝＝＝＝＝加入動作＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝寫入菜單＝＝＝＝＝＝＝＝＝＝＝
-
-  //點擊完成菜單設定，寫入菜單
   async function compeleteTrainingSetting() {
     try {
       if (totalWeight !== 0.0 && choiceAction.length > 0) {
-        if (showHistory.docID) {
+        if (showHistory && showHistory.docID) {
           const docRef = await doc(db, 'users', uid, 'trainingTables', showHistory.docID);
           const data = {
             docID: docRef.id,
@@ -392,18 +276,17 @@ const Training = () => {
             actions: choiceAction,
           };
           await updateDoc(docRef, data);
-          setOpenTrainingOne(false);
-          setOpenTrainingTwo(false);
+          setOpenTrainingPage(0);
           setOpenTrainingInput(false);
           setChoiceAction([]);
-          setTotalWeight(0);
           sendEmail();
-          setShowHistoryBackground(false);
           alertPop();
           setContent('完成設定並寄信通知');
           setTitle('');
           setDate('');
           setDescription('');
+          setVideoShow(false);
+          setPlaying(null);
         } else {
           const docRef = doc(collection(db, 'users', uid, 'trainingTables'));
           const data = {
@@ -419,18 +302,17 @@ const Training = () => {
             actions: choiceAction,
           };
           await setDoc(docRef, data);
-          setOpenTrainingOne(false);
-          setOpenTrainingTwo(false);
+          setOpenTrainingPage(0);
           setOpenTrainingInput(false);
           setChoiceAction([]);
-          setTotalWeight(0);
           sendEmail();
-          setShowHistoryBackground(false);
           alertPop();
           setContent('完成設定並寄信通知');
           setTitle('');
           setDate('');
           setDescription('');
+          setVideoShow(false);
+          setPlaying(null);
         }
       } else if (!choiceAction.length > 0) {
         alertPop();
@@ -444,18 +326,13 @@ const Training = () => {
     }
   }
 
-  // ＝＝＝＝＝＝＝＝＝＝寫入菜單＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝即時抓出每筆菜單資料＝＝＝＝＝＝＝＝＝＝＝
-
-  //第一次進來第一頁，點到哪頁跳哪頁
   useEffect(() => {
-    if (isLoggedIn == false) {
+    if (isLoggedIn === false) {
       setTrainingData([]);
     } else {
       async function getTrainingTables() {
-        if (currentPage == 1) {
-          setSkeleton(true);
+        setSkeleton(true);
+        if (currentPage === 1) {
           const first = await query(
             collection(db, 'users', uid, 'trainingTables'),
             orderBy('trainingDate', 'desc'),
@@ -472,7 +349,6 @@ const Training = () => {
             setSkeleton(false);
           }, 1000);
         } else {
-          setSkeleton(true);
           const before = await query(
             collection(db, 'users', uid, 'trainingTables'),
             orderBy('trainingDate', 'desc'),
@@ -502,9 +378,31 @@ const Training = () => {
     }
   }, [isLoggedIn, currentPage]);
 
-  //知道現在要分幾頁
+  // let pagination;
+  // useEffect(() => {
+  //   if (isLoggedIn === false) {
+  //     pagination = [];
+  //   } else {
+  //     async function getTrainingTablesPage() {
+  //       const docRef = collection(db, 'users', uid, 'trainingTables');
+  //       const querySnapshot = await getDocs(docRef);
+  //       console.log(querySnapshot.size);
+  //       const arr = [];
+  //       for (let i = 1; i <= Math.ceil(querySnapshot.size / 8); i++) {
+  //         arr.push(i);
+  //       }
+  //       pagination = arr;
+  //       console.log(pagination);
+  //       return pagination;
+  //     }
+  //     getTrainingTablesPage();
+  //   }
+  // }, [isLoggedIn]);
+
+  // console.log(pagination);
+
   useEffect(() => {
-    if (isLoggedIn == false) {
+    if (isLoggedIn === false) {
       setPagination([]);
     } else {
       async function getTrainingTablesPage() {
@@ -526,39 +424,19 @@ const Training = () => {
     }
   }, [isLoggedIn]);
 
-  // ＝＝＝＝＝＝＝＝＝＝即時抓出每筆菜單資料＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝點擊個別菜單打開內容＝＝＝＝＝＝＝＝＝＝＝
-
   function openHistory(index) {
+    setShowHistoryToggle(true);
     setLoading(true);
     setHistoryIndex(index);
     setShowHistory(trainingData[index]);
     setShowHistoryActions(trainingData[index].actions);
-    setShowHistoryToggle(true);
     setPickHistory(trainingData[index].docID);
     setImageList(trainingData[index].picture);
-    setShowPicture((prevShowPicture) => !prevShowPicture);
-    setShowHistoryBackground(true);
     setTimeout(() => {
       setLoading(false);
     }, 1000);
   }
 
-  // ＝＝＝＝＝＝＝＝＝＝點擊個別菜單打開內容＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝播放個別影片＝＝＝＝＝＝＝＝＝＝＝
-
-  function openVideo(e) {
-    setVideoUrl(promoteActions[e.target.id].videoURL);
-    setVideoShow(true);
-  }
-
-  // ＝＝＝＝＝＝＝＝＝＝播放個別影片＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝上傳照片、顯示照片、個別對應＝＝＝＝＝＝＝＝＝＝＝
-
-  //上傳後即時顯示
   useEffect(() => {
     if (imageUpload == null) return;
     setUploadSkeleton(true);
@@ -580,44 +458,13 @@ const Training = () => {
     });
   }, [imageUpload]);
 
-  function closeHistory() {
-    setShowHistoryToggle(false);
-    setShowHistoryBackground(false);
-  }
-
-  // ＝＝＝＝＝＝＝＝＝＝上傳照片、顯示照片、個別對應＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝寄信＝＝＝＝＝＝＝＝＝＝
-
   const form = useRef();
-
-  const sendEmail = (e) => {
-    emailjs.sendForm('service_aqtfkmw', 'template_jq89u95', form.current, 'c1RPxdcmtzncbu3Wv').then(
-      (result) => {
-        console.log(result.text);
-      },
-      (error) => {
-        console.log(error.text);
-      }
-    );
+  const sendEmail = () => {
+    emailjs.sendForm('service_aqtfkmw', 'template_jq89u95', form.current, 'c1RPxdcmtzncbu3Wv');
   };
 
-  // ＝＝＝＝＝＝＝＝＝＝寄信＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝關閉影片＝＝＝＝＝＝＝＝＝＝＝
-
-  function closeVideo() {
-    setVideoShow(false);
-    setPlaying(null);
-  }
-
-  // ＝＝＝＝＝＝＝＝＝＝關閉影片＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝喜愛動作功能＝＝＝＝＝＝＝＝＝＝＝
-
-  //把喜愛動作資料從雲端抓下來
   async function getFavoriteTrainings() {
-    if (isLoggedIn == false) {
+    if (isLoggedIn === false) {
       setFavoriteTrainings([]);
     } else {
       const docRef = query(collection(db, 'users', uid, 'favoriteTrainings'), orderBy('setDate'));
@@ -631,7 +478,6 @@ const Training = () => {
     }
   }
 
-  //map出選到的喜愛動作
   useEffect(() => {
     const showFavoriteTrainings = async () => {
       const docRef = doc(db, 'users', uid, 'favoriteTrainings', favoriteChoice);
@@ -645,7 +491,6 @@ const Training = () => {
     showFavoriteTrainings();
   }, [favoriteChoice]);
 
-  //點擊管理喜愛菜單
   async function manageFavoriteTraining() {
     if (isLoggedIn) {
       setOpenFavorite(true);
@@ -662,61 +507,35 @@ const Training = () => {
     }
   }
 
-  // ＝＝＝＝＝＝＝＝＝＝喜愛動作功能＝＝＝＝＝＝＝＝＝＝＝
-
-  // ＝＝＝＝＝＝＝＝＝＝歷史菜單編輯功能＝＝＝＝＝＝＝＝＝＝＝
-
   async function editTraining() {
     setChoiceAction(showHistory.actions);
     setTitle(showHistory.title);
     setDescription(showHistory.description);
     setDate(showHistory.trainingDate);
-    setShowHistoryToggle(false);
     setOpenTrainingInput(true);
-    setOpenTrainingOne(true);
+    setShowHistoryToggle(false);
+    setOpenTrainingPage(1);
     setImageList(showHistory.picture);
   }
 
-  // ＝＝＝＝＝＝＝＝＝＝歷史菜單編輯功能＝＝＝＝＝＝＝＝＝＝＝
-
   return (
     <>
-      <LoadingOutside $isActive={loading}>
-        <LoadingBlocks>
-          <Blocks
-            visible={true}
-            height="260"
-            width="260"
-            ariaLabel="blocks-loading"
-            wrapperStyle={{}}
-            wrapperClass="blocks-wrapper"
-          />
-        </LoadingBlocks>
-      </LoadingOutside>
+      {loading && (
+        <LoadingOutside>
+          <LoadingBlocks>
+            <Blocks
+              visible={true}
+              height="260"
+              width="260"
+              ariaLabel="blocks-loading"
+              wrapperStyle={{}}
+              wrapperClass="blocks-wrapper"
+            />
+          </LoadingBlocks>
+        </LoadingOutside>
+      )}
       {deleteAlert && (
-        <>
-          <DeleteAlertOutside>
-            <DeleteContent>
-              <DeletePic>
-                <FontAwesomeIcon icon={faTriangleExclamation} />
-              </DeletePic>
-              確定執行刪除？
-            </DeleteContent>
-            <DeleteButton>
-              <YesOutside onClick={confirmDeleteTrainingItem}>
-                <Yes>YES</Yes>
-              </YesOutside>
-              <NoOutside
-                onClick={() => {
-                  setDeleteAlert(false);
-                }}
-              >
-                <No>NO</No>
-              </NoOutside>
-            </DeleteButton>
-          </DeleteAlertOutside>
-          <DeleteBackground />
-        </>
+        <DeleteZone setDeleteAlert={setDeleteAlert} confirmDeleteTrainingItem={confirmDeleteTrainingItem} />
       )}
       <Wrapper>
         <BannerOutside>
@@ -725,21 +544,11 @@ const Training = () => {
           </Banner>
         </BannerOutside>
         <TrainingZone>
-          <TrainingButtons>
-            <AddTrainingTableOutside>
-              <AddTrainingTable
-                onClick={() => {
-                  addTraining();
-                  getFavoriteTrainings();
-                }}
-              >
-                建立菜單
-              </AddTrainingTable>
-            </AddTrainingTableOutside>
-            <ManageFavoriteTrainingOutside>
-              <ManageFavoriteTraining onClick={manageFavoriteTraining}>喜愛菜單</ManageFavoriteTraining>
-            </ManageFavoriteTrainingOutside>
-          </TrainingButtons>
+          <TrainingButtonsZone
+            addTraining={addTraining}
+            getFavoriteTrainings={getFavoriteTrainings}
+            manageFavoriteTraining={manageFavoriteTraining}
+          />
           <FavoritePage
             openFavorite={openFavorite}
             setOpenFavorite={setOpenFavorite}
@@ -752,24 +561,14 @@ const Training = () => {
             pickID={pickID}
             setPickID={setPickID}
           />
-          {skeleton && (
-            <>
-              <SkeletonPage />
-            </>
-          )}
+          {skeleton && <SkeletonPage />}
           {trainingData.length > 0 ? (
-            <>
-              {!skeleton && (
-                <>
-                  <HistoryZone trainingData={trainingData} openHistory={openHistory} />
-                </>
-              )}
-            </>
+            <>{!skeleton && <HistoryZone trainingData={trainingData} openHistory={openHistory} />}</>
           ) : (
             <NoHistory>尚未建立菜單</NoHistory>
           )}
         </TrainingZone>
-        {pagination.length > 0 ? (
+        {pagination && (
           <PaginationOutside>
             {pagination.map((item, index) => (
               <PaginationItem
@@ -777,28 +576,27 @@ const Training = () => {
                 onClick={() => {
                   setCurrentPgae(index + 1);
                 }}
-                $isActive={index + 1 == currentPage}
+                $isActive={index + 1 === currentPage}
               >
                 {item}
               </PaginationItem>
             ))}
           </PaginationOutside>
-        ) : null}
+        )}
         <OpenHistoryZone
           showHistory={showHistory}
+          setShowHistory={setShowHistory}
           openHistory={openHistory}
-          showHistoryBackground={showHistoryBackground}
-          showHistoryToggle={showHistoryToggle}
-          closeHistory={closeHistory}
           showHistoryActions={showHistoryActions}
           setShowHistoryActions={setShowHistoryActions}
+          showHistoryToggle={showHistoryToggle}
+          setShowHistoryToggle={setShowHistoryToggle}
           imageList={imageList}
           setImageList={setImageList}
           imageUpload={imageUpload}
           setImageUpload={setImageUpload}
           completeTraining={completeTraining}
-          setOpenCompleteSetting={setOpenCompleteSetting}
-          deleteTrainingItem={deleteTrainingItem}
+          setDeleteAlert={setDeleteAlert}
           choiceAction={choiceAction}
           data={data}
           editTraining={editTraining}
@@ -807,171 +605,120 @@ const Training = () => {
         />
         {openTrainingInput && (
           <>
-            <TrainingOutside $isHide={openTrainingInput} $isActive={pageTwo}>
-              <TrainingOutsideOne $isHide={openTrainingOne}>
-                <PageOneDetail>
-                  <Close onClick={closeAddTraining}>
-                    <FontAwesomeIcon icon={faCircleXmark} />
-                  </Close>
-                  <form ref={form}>
-                    <PageOneDetailContent>
-                      <TitleInputText>
-                        <FavoriteTitle>
-                          <Title>
-                            <FaDumbbell>
-                              <FontAwesomeIcon icon={faDumbbell} />
-                            </FaDumbbell>
-                            主題
-                          </Title>
-                          <FavoriteSelectOutside
-                            onChange={(e) => setFavoriteChoice(e.target.value)}
-                            defaultValue={null}
-                          >
-                            {favoriteTrainings.length > 0 ? (
-                              <>
-                                <option disabled selected>
-                                  套用喜愛菜單
-                                </option>
-                                {favoriteTrainings.map((item, index) => (
-                                  <option index={index} value={item.docID}>
-                                    {item.title}
-                                  </option>
-                                ))}
-                              </>
-                            ) : (
-                              <option disabled selected>
-                                無喜愛菜單
-                              </option>
-                            )}
-                          </FavoriteSelectOutside>
-                        </FavoriteTitle>
-                        <TitleInputLine />
-                        <TitleInput
-                          onChange={(e) => setTitle(e.target.value)}
-                          name="to_title"
-                          value={title}
-                          maxLength={10}
-                        ></TitleInput>
-                      </TitleInputText>
-                      <TitleRemind>＊最多輸入10字</TitleRemind>
-                      <DateInputText>
-                        <DateInputTop>
-                          <FaCalendarDays>
-                            <FontAwesomeIcon icon={faCalendarDays} />
-                          </FaCalendarDays>
-                          日期
-                        </DateInputTop>
-                        <DateInputLine />
-                        <DateInput
-                          type="date"
-                          onChange={(e) => setDate(e.target.value)}
-                          name="to_date"
-                          value={date}
-                        ></DateInput>
-                      </DateInputText>
-                      <DescriptionText>
-                        <DescriptionTop>
-                          <FaHandPointUp>
-                            <FontAwesomeIcon icon={faHandPointUp} />
-                          </FaHandPointUp>
-                          本次訓練重點
-                        </DescriptionTop>
-                        <DescriptionLine />
-                        <DescriptionInput
-                          name="to_description"
-                          onChange={(e) => setDescription(e.target.value)}
-                          value={description}
-                          maxLength={30}
-                        ></DescriptionInput>
-                      </DescriptionText>
-                      <DescriptionRemind>＊最多輸入30字</DescriptionRemind>
-                      <ToName name="to_name" defaultValue={displayName}></ToName>
-                      <ToEmail name="to_email" defaultValue={email}></ToEmail>
-                    </PageOneDetailContent>
-                  </form>
-                </PageOneDetail>
-                <PageOnePicOutside>
-                  <PageOnePic />
-                </PageOnePicOutside>
-                <TurnOutside>
-                  <TurnRight onClick={getPageTwo}>
-                    <FontAwesomeIcon icon={faCircleArrowRight} />
-                  </TurnRight>
-                </TurnOutside>
-              </TrainingOutsideOne>
-              <TrainingOutsideTwo $isHide={openTrainingTwo}>
+            <TrainingOutsideOne $isActive={openTrainingPage === 1}>
+              <PageOneDetail>
                 <Close onClick={closeAddTraining}>
                   <FontAwesomeIcon icon={faCircleXmark} />
                 </Close>
-                <ActionText>
-                  <ActionTop>
-                    <ActionPicOutside>
-                      <ActionPic src={logo} />
-                    </ActionPicOutside>
-                    <ActionTitle> 加入菜單動作</ActionTitle>
-                  </ActionTop>
-                  <ActionLine />
-                </ActionText>
-                <ActionOutside>
-                  <ChoiceActionOutsideZone
-                    choiceAction={choiceAction}
-                    setChoiceAction={setChoiceAction}
-                    deleteItem={deleteItem}
-                    totalWeight={totalWeight}
-                    choiceWeight={choiceWeight}
-                    setChoiceWeight={setChoiceWeight}
-                    choiceTimes={choiceTimes}
-                    setChoiceTimes={setChoiceTimes}
-                  />
-                  <PromoteActionOutsideZone
-                    setPart={setPart}
-                    promoteActions={promoteActions}
-                    addActionItem={addActionItem}
-                    openVideo={openVideo}
-                    playing={playing}
-                    setPlaying={setPlaying}
-                  />
-                </ActionOutside>
-                <TrainingOutsideTwoBottom>
-                  <CalculationShowZone
-                    choiceAction={choiceAction}
-                    data={data}
-                    dataNull={dataNull}
-                    compeleteTrainingSetting={compeleteTrainingSetting}
-                    sendEmail={sendEmail}
-                  />
-                  {videoShow ? (
-                    <VideoZone $isHide={videoShow}>
-                      <CloseVideo onClick={closeVideo}>
-                        <FontAwesomeIcon icon={faCircleXmark} />
-                      </CloseVideo>
-                      <video autoPlay loop width="100%" controls src={videoUrl}></video>
-                    </VideoZone>
-                  ) : (
-                    <NoVideo>
-                      <NoVideoTitle>範例影片播放區</NoVideoTitle>
-                    </NoVideo>
-                  )}
-                </TrainingOutsideTwoBottom>
-                <CompeleteTrainingSettingOutside>
-                  <CompeleteTrainingSetting
-                    type="button"
-                    value="submit"
-                    onClick={() => {
-                      compeleteTrainingSetting();
-                    }}
-                  >
-                    完成菜單設定
-                  </CompeleteTrainingSetting>
-                </CompeleteTrainingSettingOutside>
-                <TurnOutside>
-                  <TurnLeft onClick={getPageOne}>
-                    <FontAwesomeIcon icon={faCircleArrowLeft} />
-                  </TurnLeft>
-                </TurnOutside>
+                <form ref={form}>
+                  <PageOneDetailContent>
+                    <TitleInputText>
+                      <FavoriteTitle>
+                        <Title>
+                          <FaDumbbell>
+                            <FontAwesomeIcon icon={faDumbbell} />
+                          </FaDumbbell>
+                          主題
+                        </Title>
+                        <FavoriteSelectOutside onChange={(e) => setFavoriteChoice(e.target.value)} defaultValue={null}>
+                          {favoriteTrainings.length > 0 ? (
+                            <>
+                              <option disabled selected>
+                                套用喜愛菜單
+                              </option>
+                              {favoriteTrainings.map((item, index) => (
+                                <option index={index} value={item.docID}>
+                                  {item.title}
+                                </option>
+                              ))}
+                            </>
+                          ) : (
+                            <option disabled selected>
+                              無喜愛菜單
+                            </option>
+                          )}
+                        </FavoriteSelectOutside>
+                      </FavoriteTitle>
+                      <TitleInputLine />
+                      <TitleInput
+                        onChange={(e) => setTitle(e.target.value)}
+                        name="to_title"
+                        value={title}
+                        maxLength={10}
+                      ></TitleInput>
+                    </TitleInputText>
+                    <TitleRemind>＊最多輸入10字</TitleRemind>
+                    <DateInputText>
+                      <DateInputTop>
+                        <FaCalendarDays>
+                          <FontAwesomeIcon icon={faCalendarDays} />
+                        </FaCalendarDays>
+                        日期
+                      </DateInputTop>
+                      <DateInputLine />
+                      <DateInput
+                        type="date"
+                        onChange={(e) => setDate(e.target.value)}
+                        name="to_date"
+                        value={date}
+                      ></DateInput>
+                    </DateInputText>
+                    <DescriptionText>
+                      <DescriptionTop>
+                        <FaHandPointUp>
+                          <FontAwesomeIcon icon={faHandPointUp} />
+                        </FaHandPointUp>
+                        本次訓練重點
+                      </DescriptionTop>
+                      <DescriptionLine />
+                      <DescriptionInput
+                        name="to_description"
+                        onChange={(e) => setDescription(e.target.value)}
+                        value={description}
+                        maxLength={30}
+                      ></DescriptionInput>
+                    </DescriptionText>
+                    <DescriptionRemind>＊最多輸入30字</DescriptionRemind>
+                    <ToName name="to_name" defaultValue={displayName}></ToName>
+                    <ToEmail name="to_email" defaultValue={email}></ToEmail>
+                  </PageOneDetailContent>
+                </form>
+              </PageOneDetail>
+              <PageOnePicOutside>
+                <PageOnePic />
+              </PageOnePicOutside>
+              <TurnOutside>
+                <TurnRight onClick={getPageTwo}>
+                  <FontAwesomeIcon icon={faCircleArrowRight} />
+                </TurnRight>
+              </TurnOutside>
+            </TrainingOutsideOne>
+            {openTrainingPage === 2 && (
+              <TrainingOutsideTwo>
+                <TrainingTwo
+                  closeAddTraining={closeAddTraining}
+                  choiceAction={choiceAction}
+                  setChoiceAction={setChoiceAction}
+                  deleteItem={deleteItem}
+                  totalWeight={totalWeight}
+                  setPart={setPart}
+                  promoteActions={promoteActions}
+                  addActionItem={addActionItem}
+                  playing={playing}
+                  setPlaying={setPlaying}
+                  setVideoShow={setVideoShow}
+                  videoShow={videoShow}
+                  setVideoUrl={setVideoUrl}
+                  videoUrl={videoUrl}
+                  data={data}
+                  dataNull={dataNull}
+                  compeleteTrainingSetting={compeleteTrainingSetting}
+                  sendEmail={sendEmail}
+                  setOpenTrainingPage={setOpenTrainingPage}
+                />
               </TrainingOutsideTwo>
-            </TrainingOutside>
-            <TrainingBackground $isHide={showHistoryBackground} />
+            )}
+            <Background />
           </>
         )}
       </Wrapper>
@@ -980,6 +727,21 @@ const Training = () => {
 };
 
 export default Training;
+
+const Close = styled.div`
+  cursor: pointer;
+  width: 30px;
+  position: absolute;
+  right: 20px;
+  top: 10px;
+  scale: 1;
+  transition: 0.3s;
+  font-size: 30px;
+  color: #c14e4f;
+  &:hover {
+    scale: 1.2;
+  }
+`;
 
 const PaginationOutside = styled.div`
   display: flex;
@@ -1007,7 +769,7 @@ const LoadingOutside = styled.div`
   background: #475260;
   height: 100%;
   width: 100%;
-  display: ${(props) => (props.$isActive ? 'block' : 'none')};
+  display: block;
 `;
 
 const LoadingBlocks = styled.div`
@@ -1018,121 +780,6 @@ const LoadingBlocks = styled.div`
   transform: translate(-50%, -50%);
 `;
 
-const DeleteAlertOutside = styled.div`
-  display: flex;
-  padding: 15px;
-  width: 400px;
-  background: #475260;
-  border: 5px solid #74c6cc;
-  border-radius: 20px;
-  position: absolute;
-  top: 35%;
-  left: calc(50% - 200px);
-  z-index: 100;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  animation-name: deletefadein;
-  animation-duration: 0.8s;
-  @keyframes deletefadein {
-    0% {
-      transform: translateY(-10%);
-      opacity: 0%;
-    }
-    100% {
-      transform: translateY(0%);
-      opacity: 100%;
-    }
-  }
-  @media screen and (max-width: 767px) {
-    padding: 10px;
-    width: 300px;
-    left: calc(50% - 150px);
-  }
-`;
-
-const DeletePic = styled.div`
-  color: #ffd700;
-  font-size: 40px;
-  margin-right: 20px;
-`;
-
-const DeleteContent = styled.div`
-  display: flex;
-  align-items: center;
-  color: white;
-  font-size: 30px;
-  margin-top: 10px;
-  letter-spacing: 6px;
-  @media screen and (max-width: 767px) {
-    margin-top: 0px;
-    font-size: 25px;
-    letter-spacing: 4px;
-  }
-`;
-const DeleteButton = styled.div`
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  margin-top: 15px;
-  margin-bottom: 10px;
-`;
-
-const YesOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 50px;
-  color: black;
-  cursor: pointer;
-  border-radius: 10px;
-  margin-right: 30px;
-  &:hover {
-    background: #c14e4f;
-    color: black;
-  }
-`;
-
-const Yes = styled.div`
-  padding: 3px;
-  font-size: 20px;
-  letter-spacing: 2px;
-  font-weight: 600;
-`;
-
-const NoOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 50px;
-  color: black;
-  cursor: pointer;
-  border-radius: 10px;
-  &:hover {
-    background: white;
-    color: black;
-  }
-`;
-
-const No = styled.div`
-  padding: 3px;
-  font-size: 20px;
-  letter-spacing: 2px;
-  font-weight: 600;
-`;
-
-const DeleteBackground = styled.div`
-  background: black;
-  top: 0;
-  opacity: 50%;
-  z-index: 60;
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-`;
-
 const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -1140,21 +787,6 @@ const Wrapper = styled.div`
   font-size: 20px;
   position: relative;
   padding-top: 90px;
-`;
-
-const Close = styled.div`
-  cursor: pointer;
-  width: 30px;
-  position: absolute;
-  right: 20px;
-  top: 10px;
-  scale: 1;
-  transition: 0.3s;
-  font-size: 30px;
-  color: #c14e4f;
-  &:hover {
-    scale: 1.2;
-  }
 `;
 
 const BannerOutside = styled.div`
@@ -1210,97 +842,6 @@ const TrainingZone = styled.div`
   flex-direction: column;
 `;
 
-const TrainingButtons = styled.div`
-  display: flex;
-  max-width: 1200px;
-  margin: 0 auto;
-`;
-
-const AddTrainingTableOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 180px;
-  margin: 40px 20px 40px auto;
-  color: black;
-  cursor: pointer;
-  transition: ease-in-out 0.2s;
-  animation-duration: 2.5s;
-  animation-iteration-count: infinite;
-  &:hover {
-    background: white;
-    color: black;
-  }
-  @keyframes light {
-    0% {
-      box-shadow: 0px 0px 0px white;
-    }
-    50% {
-      box-shadow: 0px 0px 20px white;
-    }
-    100% {
-      box-shadow: 0px 0px 0px white;
-    }
-  }
-  @media screen and (max-width: 767px) {
-    width: 120px;
-    margin: 40px 20px 40px 0px;
-  }
-`;
-
-const AddTrainingTable = styled.div`
-  padding: 10px;
-  font-size: 25px;
-  letter-spacing: 2px;
-  font-weight: 600;
-  @media screen and (max-width: 767px) {
-    font-size: 18px;
-  }
-`;
-
-const ManageFavoriteTrainingOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 180px;
-  margin: 40px auto 40px 20px;
-  color: black;
-  cursor: pointer;
-  transition: ease-in-out 0.2s;
-  animation-duration: 2.5s;
-  animation-iteration-count: infinite;
-  &:hover {
-    background: white;
-    color: black;
-  }
-  @keyframes light {
-    0% {
-      box-shadow: 0px 0px 0px white;
-    }
-    50% {
-      box-shadow: 0px 0px 20px white;
-    }
-    100% {
-      box-shadow: 0px 0px 0px white;
-    }
-  }
-  @media screen and (max-width: 767px) {
-    width: 120px;
-  }
-`;
-
-const ManageFavoriteTraining = styled.div`
-  padding: 10px;
-  font-size: 25px;
-  letter-spacing: 2px;
-  font-weight: 600;
-  @media screen and (max-width: 767px) {
-    font-size: 18px;
-  }
-`;
-
 const NoHistory = styled.div`
   max-width: 1200px;
   margin: auto;
@@ -1314,16 +855,14 @@ const NoHistory = styled.div`
   }
 `;
 
-// ＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝第一頁＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝
-
-const TrainingOutside = styled.div`
+const TrainingOutsideOne = styled.div`
   position: absolute;
-  left: ${(props) => (props.$isActive ? 'calc(50% - 500px)' : 'calc(50% - 280px)')};
-  top: ${(props) => (props.$isActive ? '7%' : '10%')};
+  left: calc(50% - 280px);
+  top: 10%;
   z-index: 20;
-  display: ${(props) => (props.$isHide ? 'block' : 'none')};
+  display: ${(props) => (props.$isActive ? 'bloock' : 'none')};
   background: #475260;
-  max-width: ${(props) => (props.$isActive ? '1000px' : '700px')};
+  max-width: 700px;
   margin-bottom: 100px;
   margin-top: 100px;
   color: white;
@@ -1341,23 +880,59 @@ const TrainingOutside = styled.div`
     }
   }
   @media screen and (max-width: 1279px) {
-    left: ${(props) => (props.$isActive ? 'calc(50% - 350px)' : 'calc(50% - 280px)')};
-    top: ${(props) => (props.$isActive ? '4%' : '5%')};
+    left: calc(50% - 280px);
+    top: 5%;
   }
   @media screen and (max-width: 767px) {
-    left: ${(props) => (props.$isActive ? 'calc(50% - 275px)' : 'calc(50% - 220px)')};
-    top: ${(props) => (props.$isActive ? '2%' : '2%')};
+    left: calc(50% - 220px);
+    top: 2%;
   }
   @media screen and (max-width: 575px) {
-    left: ${(props) => (props.$isActive ? 'calc(50% - 165px)' : 'calc(50% - 220px)')};
+    left: calc(50% - 220px);
   }
   @media screen and (max-width: 500px) {
-    left: ${(props) => (props.$isActive ? 'calc(50% - 165px)' : 'calc(50% - 170px)')};
+    left: calc(50% - 170px);
   }
 `;
 
-const TrainingOutsideOne = styled.div`
-  display: ${(props) => (props.$isHide ? 'block' : 'none')};
+const TrainingOutsideTwo = styled.div`
+  position: absolute;
+  left: calc(50% - 500px);
+  top: 7%;
+  z-index: 20;
+  display: block;
+  background: #475260;
+  max-width: 1000px;
+  margin-bottom: 100px;
+  margin-top: 100px;
+  color: white;
+  border-top: 0.5rem solid #74c6cc;
+  animation-name: favoritefadein;
+  animation-duration: 0.5s;
+  @keyframes favoritefadein {
+    0% {
+      transform: translateY(-2%);
+      opacity: 0%;
+    }
+    100% {
+      transform: translateY(0%);
+      opacity: 100%;
+    }
+  }
+  @media screen and (max-width: 1279px) {
+    left: calc(50% - 350px);
+    top: 4%;
+  }
+  @media screen and (max-width: 767px) {
+    left: calc(50% - 275px);
+    top: 2%;
+  }
+  @media screen and (max-width: 575px) {
+    left: calc(50% - 165px);
+  }
+  @media screen and (max-width: 500px) {
+    left: calc(50% - 165px);
+  }
 `;
 
 const PageOneDetail = styled.div`
@@ -1573,162 +1148,6 @@ const TurnRight = styled.div`
   }
 `;
 
-// ＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝第一頁＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝
-
-// ＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝第二頁＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝＝
-
-const TrainingOutsideTwo = styled.div`
-  display: ${(props) => (props.$isHide ? 'block;' : 'none;')};
-`;
-
-const ActionText = styled.div`
-  padding-top: 30px;
-  padding-left: 30px;
-  padding-right: 30px;
-  padding-bottom: 20px;
-  color: #74c6cc;
-  font-size: 24px;
-  font-weight: 600;
-  letter-spacing: 3px;
-
-  @media screen and (max-width: 767px) {
-    padding-bottom: 0px;
-  }
-`;
-
-const ActionTop = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const ActionPicOutside = styled.div``;
-
-const ActionPic = styled.img`
-  width: 70px;
-  margin-right: 15px;
-  padding-top: 10px;
-`;
-
-const ActionTitle = styled.div``;
-
-const ActionLine = styled.div`
-  border-bottom: 2px solid #74c6cc;
-  margin-top: 15px;
-  margin-bottom: 10px;
-`;
-
-const ActionOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 1000px;
-  @media screen and (max-width: 1279px) {
-    flex-direction: column;
-    width: 700px;
-  }
-  @media screen and (max-width: 767px) {
-    flex-direction: column;
-    width: 550px;
-  }
-  @media screen and (max-width: 575px) {
-    flex-direction: column;
-    margin: 0 auto;
-    width: 320px;
-  }
-`;
-
-const CloseVideo = styled.div`
-  cursor: pointer;
-  width: 30px;
-  height: 30px;
-  z-index: 90;
-  position: absolute;
-  scale: 1;
-  right: 50px;
-  color: #c14e4f;
-  font-size: 30px;
-  &:hover {
-    scale: 1.2;
-  }
-  @media screen and (max-width: 767px) {
-    top: 10px;
-    right: 10px;
-  }
-  @media screen and (max-width: 575px) {
-    top: 10px;
-    right: 10px;
-  }
-`;
-
-const VideoZone = styled.div`
-  display: ${(props) => (props.$isHide ? 'block;' : 'none;')};
-  z-index: 65;
-  padding: 0px 40px;
-  width: 550px;
-  position: relative;
-  @media screen and (max-width: 1279px) {
-    margin-top: 40px;
-  }
-  @media screen and (max-width: 767px) {
-    width: 350px;
-    padding: 0px 0px;
-  }
-  @media screen and (max-width: 575px) {
-    width: 300px;
-    padding: 0px 0px;
-    margin-right: 0px;
-  }
-`;
-
-const NoVideoTitle = styled.div`
-  width: 200px;
-  text-align: center;
-`;
-
-const NoVideo = styled.div`
-  width: 450px;
-  height: 253.13px;
-  border: 1px solid white;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0px 40px;
-  @media screen and (max-width: 1279px) {
-    margin-top: 40px;
-  }
-  @media screen and (max-width: 767px) {
-    width: 350px;
-  }
-  @media screen and (max-width: 575px) {
-    width: 250px;
-  }
-`;
-
-const TrainingOutsideTwoBottom = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  @media screen and (max-width: 1279px) {
-    flex-direction: column-reverse;
-  }
-`;
-
-const TurnLeft = styled.div`
-  cursor: pointer;
-  margin-right: auto;
-  margin-left: 13px;
-  margin-top: 10px;
-  margin-bottom: 10px;
-  font-size: 30px;
-  &:hover {
-    color: #74c6cc;
-  }
-`;
-
-const TurnOutside = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
 const ToEmail = styled.input`
   display: none;
 `;
@@ -1737,7 +1156,12 @@ const ToName = styled.input`
   display: none;
 `;
 
-const TrainingBackground = styled.div`
+const TurnOutside = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+const Background = styled.div`
   background: black;
   top: 0;
   opacity: 50%;
@@ -1745,31 +1169,5 @@ const TrainingBackground = styled.div`
   position: fixed;
   width: 100vw;
   height: 100vh;
-  display: ${(props) => (props.$isHide ? 'block' : 'none')};
-`;
-
-const CompeleteTrainingSettingOutside = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #74c6cc;
-  width: 140px;
-  margin-top: 15px;
-  margin-bottom: 10px;
-  margin-left: auto;
-  margin-right: auto;
-  color: black;
-  cursor: pointer;
-  &:hover {
-    background: white;
-    color: black;
-  }
-`;
-
-const CompeleteTrainingSetting = styled.div`
-  cursor: pointer;
-  padding: 8px;
-  font-size: 18px;
-  letter-spacing: 1.2px;
-  font-weight: 600;
+  display: block;
 `;
